conditions added on deleted users view

+++ app/Http/Controllers/HomeController.php
@@ -29,21 +29,33 @@ class HomeController extends Controller
 
     public function viewDeletedUsers()
     {
-        $users = User::onlyTrashed()->get();
+        if (auth()->user()->id == 1) {
+            $users = User::onlyTrashed()->get();
+        } else {
+            return redirect()->back();
+        }
 
         return view('deleted_users')->with('users', $users);
     }
 
     public function restoreUser($user_id)
     {
-        User::withTrashed()->find($user_id)->restore();
+        if (auth()->user()->id == 1) {
+            User::withTrashed()->find($user_id)->restore();
+        } else {
+            return redirect()->back();
+        }
 
         return redirect()->back();
     }
 
     public function hardDeleteUser($user_id)
     {
-        User::withTrashed()->find($user_id)->forceDelete();
+        if (auth()->user()->id == 1) {
+            User::withTrashed()->find($user_id)->forceDelete();
+        } else {
+            return redirect()->back();
+        }
 
         return redirect()->back();
     }
